+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Added
 ### Fixed
+- Fixes some translations and phrases in the audio settings.
 
 ## [7.0.2] - 2025-10-01
 ### Added